throttle: rework the Throttle API
Instead of capturing the function in the Throttle, pass it in via the Call
method. This allows a Throttle to be used as a field on a struct without a
delicate dance to allocate and initialize the field. A zero Throttle is now
ready for use, with the usual caveats about copying.