msync: rename Flag to Relay
This is a breaking change, but no uses found anywhere on GitHub. Maybe a better
approach would be to delete it entirely. But for now, let's keep it.

This effectively reverts commit c2f95118fe3b. I didn't like the name "Handoff",
but "Flag" was no better. I think Relay is an improvement.